data/requirements.txt
appnope==0.1.0
beautifulsoup4==4.4.1
boto==2.39.0
bz2file==0.98
CacheControl==0.11.6
contextlib2==0.5.1
cycler==0.10.0
decorator==4.0.9
funcy==1.6
future==0.15.2
gensim==0.12.4
gnureadline==6.3.3
httpretty==0.8.10
ipykernel==4.3.1
ipython==4.1.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
Jinja2==2.8
joblib==0.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
lockfile==0.12.2
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.2
natsort==4.0.4
nbconvert==4.1.0
nbformat==4.0.1
nltk==3.2
nose==1.3.7
notebook==4.1.0
numexpr==2.5
numpy==1.10.4
pandas==0.17.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
py==1.4.31
Pygments==2.1.3
pyLDAvis==1.4.1
pyparsing==2.1.0
pytest==2.8.7
python-dateutil==2.4.2
pytz==2015.7
pyzmq==15.2.0
qtconsole==4.2.0
requests==2.8.1
scikit-bio==0.4.2
scikit-learn==0.17.1
scipy==0.17.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
smart-open==1.3.2
terminado==0.6
tornado==4.3
traitlets==4.1.0
